p5odds.odd bug fix (#2629)
* Bug fix:
Constraint was looking for _all_ descendant text nodes of eg element,
not just child text nodes. That means that if there is any phrase-level
encoding in an eg element there is more than one text node that has (for
example) position()=last(), and the XPath fails because only one item is
allowed as the first argument to matches().

* Better fix:
Per suggestion of @martindholmes, instead of trying to select the correct
descendant text node(s), just examine the string value of the entire eg
element as a whole. The string() function call is not technically necessary
(because matches() wants a string there, it would happen automagically),
but its use makes what we are doing explicit and obvious.